Extract http module in pheidippides_utils
