+++ lib/pheidippides-utils/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod utils;
 pub mod async_utils;
 pub mod serde;
+pub mod http;
