+++ lib/pheidippides-utils/src/utils.rs
@@ -1,9 +1,4 @@
-use std::collections::HashMap;
 use std::hash::Hash;
-
-use anyhow::Result;
-
-pub type Header = (CaseInsensitiveString, String);
 
 #[derive(Debug, PartialEq, Eq, Hash)]
 pub struct CaseInsensitiveString(String);
@@ -22,28 +17,4 @@ impl std::fmt::Display for CaseInsensitiveString {
 
 pub fn log_internal_error(error: impl std::fmt::Display) {
     eprintln!("SERVER ERROR: {:#}", error);
-}
-
-pub enum CookieParsingError {
-    IncorrectHeader,
-}
-
-pub fn get_cookies_hashmap(
-    headers: &HashMap<CaseInsensitiveString, String>,
-) -> Result<HashMap<String, String>, CookieParsingError> {
-    let mut res = HashMap::new();
-    if let Some(cookie_list) = headers.get(&"Cookie".into()) {
-        for cookie in cookie_list.split("; ") {
-            let (key, value) = match cookie.split_once('=') {
-                Some(key_value) => key_value,
-                None => return Err(CookieParsingError::IncorrectHeader),
-            };
-            res.insert(key.into(), value.into());
-        }
-    }
-    Ok(res)
-}
-
-pub fn header_set_cookie(key: &str, value: &str) -> Header {
-    ("Set-Cookie".into(), format!("{key}={value}"))
 }
